+++ include/NullGameEngine/NullGameEngine.hpp
@@ -2,7 +2,6 @@
 
 namespace null {
     class Component;
-    class Script;
     class GameObject;
     class Scene;
     class SceneLoader;
