+++ include/NullGameEngine/SceneLoader.hpp
@@ -1,0 +1,6 @@
+#include <NullGameEngine.hpp>
+
+class SceneLoader {
+
+};
+
