+++ include/NullGameEngine/Script.hpp
@@ -1,12 +1,8 @@
 #pragma once
 
 #include <NullGameEngine.hpp>
-#include <Component.hpp>
 
 namespace null {
-    class Script : public Component{
-    public:
-        virtual void start();
-        virtual void update();
-    };
+    // scripts are essentially additional components
+    using Script = Component;
 }
