+++ src/NullGameEngine/Component.cpp
@@ -1,0 +1,8 @@
+#include <Component.hpp>
+
+namespace null {
+
+
+
+}
+
