+++ src/NullGameEngine/SceneLoader.cpp
@@ -1,0 +1,6 @@
+#include <SceneLoader.hpp>
+
+namespace null {
+
+}
+
